requirements.txt
altair==4.2.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.6.2
boto3==1.26.165
botocore==1.29.165
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.0.3
cssutils==2.7.1
cv==1.0.0
cycler==0.11.0
debugpy==1.6.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.4
django-cors-headers==4.1.0
django-crispy-forms==1.14.0
django-resources==1.0a1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
entrypoints==0.4
executing==0.8.3
factory-boy==3.2.1
Faker==15.3.4
fastjsonschema==2.15.3
filelock==3.12.0
Flask==2.0.2
fonttools==4.28.1
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
imutils==0.5.4
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
jsmin==3.0.1
jsonschema==4.6.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.3.2
mailchimp-transactional==1.0.50
markdown-it-py==2.2.0
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.21.4
opencv-python==4.7.0.72
packaging==21.2
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pipenv==2023.4.20
platformdirs==3.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.3
psutil==5.9.1
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
requests==2.28.2
response==0.5.0
rich==13.4.1
s3transfer==0.6.1
schedule==1.1.0
scipy==1.10.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
soupsieve==2.3.2.post1
SpeechRecognition==3.10.0
sqlparse==0.4.3
stack-data==0.3.0
streamlit==1.22.0
tenacity==8.2.2
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
typing_extensions==4.6.2
tzlocal==5.0.1
urllib3==1.26.14
uuid==1.30
validators==0.20.0
virtualenv==20.22.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==6.5.0
widgetsnbextension==3.6.1
zipp==3.15.0
